feat: use experimental app directory
