+++ app/page.tsx
@@ -1,0 +1,14 @@
+import styles from '../styles/Home.module.css'
+
+export default function Home() {
+    return (
+        <div className={styles.container}>
+
+            <main className={styles.main}>
+                <p className={styles.description}>
+                    <code className={styles.code}>welcome</code>
+                </p>
+            </main>
+        </div>
+    )
+}
