--- pages/index.tsx
@@ -1,20 +1,0 @@
-import Head from 'next/head'
-import styles from '../styles/Home.module.css'
-
-export default function Home() {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>DeFi Spartan&#39;s Homepage</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <p className={styles.description}>
-          <code className={styles.code}>welcome</code>
-        </p>
-      </main>
-    </div>
-  )
-}
